GF-4[feat]: Room Create Dto 분리
